+++ src/BaseRecipeInterface.php
@@ -24,9 +24,9 @@ declare(strict_types=1);
 
 namespace Teknoo\Recipe;
 
+use Teknoo\Immutable\ImmutableInterface;
+
 /**
- * Interface to define a Cookbook. A Cookbook is a factory to write recipe.
- *
  * @copyright   Copyright (c) 2009-2020 Richard Déloge (richarddeloge@gmail.com)
  *
  * @link        http://teknoo.software/recipe Project website
@@ -34,10 +34,30 @@ namespace Teknoo\Recipe;
  * @license     http://teknoo.software/license/mit         MIT License
  * @author      Richard Déloge <richarddeloge@gmail.com>
  */
-interface CookbookInterface extends BaseRecipeInterface
+interface BaseRecipeInterface
 {
     /**
-     * To transform a recipe (can be prefilled with additional step), business rule
+     * To train a chef about this recipe.
+     *
+     * @param ChefInterface $chef
+     * @return RecipeInterface
      */
-    public function fill(RecipeInterface $recipe): CookbookInterface;
+    public function train(ChefInterface $chef): BaseRecipeInterface;
+
+    /**
+     * To prepare the work plan of the chef before start the cooking.
+     *
+     * @param array<string, mixed> $workPlan
+     * @param ChefInterface $chef
+     * @return RecipeInterface
+     */
+    public function prepare(array &$workPlan, ChefInterface $chef): BaseRecipeInterface;
+
+    /**
+     * To validate the result of the cooking via the dish defined via the method "given".
+     *
+     * @param mixed $value
+     * @return RecipeInterface
+     */
+    public function validate($value): BaseRecipeInterface;
 }