+++ src/Bowl/Bowl.php
@@ -57,7 +57,7 @@ class Bowl implements BowlInterface
 
     /**
      * @param callable $callable
-     * @param array<string, string> $mapping
+     * @param array<string, string|string[]> $mapping
      * @param string $name
      */
     public function __construct(callable $callable, array $mapping, string $name = '')