+++ tests/AbstractChefTest.php
@@ -22,6 +22,7 @@
 
 namespace Teknoo\Tests\Recipe;
 
+use Teknoo\Recipe\BaseRecipeInterface;
 use Teknoo\Recipe\Bowl\BowlInterface;
 use Teknoo\Recipe\ChefInterface;
 use PHPUnit\Framework\TestCase;
@@ -73,6 +74,19 @@ abstract class AbstractChefTest extends TestCase
         );
     }
 
+    public function testReadWithBaseRecipe()
+    {
+        $recipe = $this->createMock(BaseRecipeInterface::class);
+        $recipe->expects(self::once())
+            ->method('train')
+            ->willReturnSelf();
+
+        self::assertInstanceOf(
+            ChefInterface::class,
+            $this->buildChef()->read($recipe)
+        );
+    }
+
     public function testExceptionOnReserveAndBeginWithBadRecipe()
     {
         $this->expectException(\TypeError::class);
@@ -87,6 +101,57 @@ abstract class AbstractChefTest extends TestCase
             ->willReturnSelf();
 
         $subRecipe = $this->createMock(RecipeInterface::class);
+        $subRecipe->expects(self::once())
+            ->method('train')
+            ->willReturnCallback(function (ChefInterface $chef) use ($subRecipe) {
+                $chef->followSteps(['substep' => $this->createMock(BowlInterface::class)]);
+
+                return $subRecipe;
+            });
+
+        $chef = $this->buildChef();
+        self::assertInstanceOf(
+            ChefInterface::class,
+            $chef->read($mainRecipe)
+        );
+
+        $step = $this->createMock(BowlInterface::class);
+        $step->expects(self::once())
+            ->method('execute')
+            ->willReturnCallback(function (ChefInterface $chef) use ($step, $subRecipe) {
+                self::assertInstanceOf(
+                    ChefInterface::class,
+                    $subchef = $chef->reserveAndBegin($subRecipe)
+                );
+
+                self::assertNotSame(
+                    $chef,
+                    $subchef
+                );
+
+                return $step;
+            });
+
+        self::assertInstanceOf(
+            ChefInterface::class,
+            $chef->followSteps(['step' => $step])
+        );
+
+        self::assertInstanceOf(
+            ChefInterface::class,
+            $chef->process(['foo'=>'bar'])
+        );
+    }
+
+
+    public function testReserveAndBeginAvailableOnCookingWithBaseRecipe()
+    {
+        $mainRecipe = $this->createMock(RecipeInterface::class);
+        $mainRecipe->expects(self::once())
+            ->method('train')
+            ->willReturnSelf();
+
+        $subRecipe = $this->createMock(BaseRecipeInterface::class);
         $subRecipe->expects(self::once())
             ->method('train')
             ->willReturnCallback(function (ChefInterface $chef) use ($subRecipe) {
@@ -190,6 +255,13 @@ abstract class AbstractChefTest extends TestCase
     {
         $this->expectException(\Throwable::class);
         $recipe = $this->createMock(CookbookInterface::class);
+        $this->buildChef()->reserveAndBegin($recipe);
+    }
+
+    public function testReserveAndBeginOnNonTrainedChefWithBaseRecipe()
+    {
+        $this->expectException(\Throwable::class);
+        $recipe = $this->createMock(BaseRecipeInterface::class);
         $this->buildChef()->reserveAndBegin($recipe);
     }