+++ tests/Bowl/BowlClassTest.php
@@ -39,10 +39,11 @@ use Teknoo\Recipe\ChefInterface;
  */
 class BowlClassTest extends AbstractBowlTest
 {
-    public static function methodToCall(ChefInterface $chef, $bar, $foo2, \DateTime $date, $_methodName)
+    public static function methodToCall(ChefInterface $chef, $bar, $bar2, $foo2, \DateTime $date, $_methodName)
     {
         $chef->continue([
             'bar' => $bar,
+            'bar2' => $bar2,
             'foo2' => $foo2,
             'date' => $date->getTimestamp(),
             '_methodName' => $_methodName,
@@ -56,7 +57,7 @@ class BowlClassTest extends AbstractBowlTest
 
     protected function getMapping()
     {
-        return ['bar' => 'foo'];
+        return ['bar' => 'foo', 'bar2' => ['bar', 'foo']];
     }
 
     /**