+++ tests/Bowl/BowlClosureTest.php
@@ -41,9 +41,10 @@ class BowlClosureTest extends AbstractBowlTest
 {
     protected function getCallable()
     {
-        return function (ChefInterface $chef, $bar, $foo2, \DateTime $date, $_methodName) {
+        return function (ChefInterface $chef, $bar, $bar2, $foo2, \DateTime $date, $_methodName) {
             $chef->continue([
                 'bar' => $bar,
+                'bar2' => $bar,
                 'foo2' => $foo2,
                 'date' => $date->getTimestamp(),
                 '_methodName' => $_methodName,
@@ -53,7 +54,7 @@ class BowlClosureTest extends AbstractBowlTest
 
     protected function getMapping()
     {
-        return ['bar' => 'foo'];
+        return ['bar' => 'foo', 'bar2' => ['bar', 'foo']];
     }
 
     /**