+++ tests/Bowl/BowlInvokableTest.php
@@ -46,6 +46,8 @@ class BowlInvokableTest extends AbstractBowlTest
             {
                 $chef->continue([
                     'bar' => $bar,
+                    'bar2' => $bar,
+                    'bar2' => $bar,
                     'foo2' => $foo2,
                     'date' => $date->getTimestamp(),
                     '_methodName' => $_methodName,
@@ -58,7 +60,7 @@ class BowlInvokableTest extends AbstractBowlTest
 
     protected function getMapping()
     {
-        return ['bar' => 'foo'];
+        return ['bar' => 'foo', 'bar2' => ['bar', 'foo']];
     }
 
     /**