+++ tests/Bowl/BowlObjectTest.php
@@ -42,10 +42,11 @@ class BowlObjectTest extends AbstractBowlTest
     protected function getCallable()
     {
         $object = new class() {
-            public function methodToCall(ChefInterface $chef, $bar, $foo2, \DateTime $date, $_methodName)
+            public function methodToCall(ChefInterface $chef, $bar, $bar2, $foo2, \DateTime $date, $_methodName)
             {
                 $chef->continue([
                     'bar' => $bar,
+                    'bar2' => $bar2,
                     'foo2' => $foo2,
                     'date' => $date->getTimestamp(),
                     '_methodName' => $_methodName,
@@ -58,7 +59,7 @@ class BowlObjectTest extends AbstractBowlTest
 
     protected function getMapping()
     {
-        return ['bar' => 'foo'];
+        return ['bar' => 'foo', 'bar2' => ['bar', 'foo']];
     }
 
     /**